denylist: add branched, next-devel streams to denials for rawhide
F39 has been branched from rawhide and we'll soon be building a
branched stream again and also next-devel will soon be based on F39.
Let's add branched and next-devel to those denylist entries.

Don't apply this to the `toolbox` entry because that failure is only
because `rawhide` is now F40.