overlay.d/05core: drop Requires on root device for coreos-rootflags.service
We need to drop this Requires for a subtle reason.

In this unit we do have ConditionPathExists=!/run/ostree-live
so we'll never run on a "live" boot, but those Conditions won't
get evaluated until all other dependencies are satisfied. On a
live boot there won't be a /dev/disk/by-label/root device because
the root filesystem is coming from the ISO or PXE so the
coreos-rootflags.service would wait forever for that device to
exist and eventually fail.