transposefs: Only autosave-xfs for much larger filesystems
The change in https://github.com/coreos/fedora-coreos-config/pull/2320
has been very problematic for OpenShift because our default node
configuration is *always* over the threshold, and that causes
significant latency on instance provisioning.

Experimentally bumping to 400 allocation groups, which is about 700GiB.
This is comfortably about the default OpenShift node root disk sizes,
and returns us to the prior status quo.

While we're here, rework the logging a bit so that we *always*
log the `agcount` for debugging purposes.

Also:

- Only log to stdout for normal conditions
- Include the name of the systemd unit in the test description
  so we can cross-reference
- tests: Hoist the expected agcount of 4 to a common variable