kola secex tests: Increase test timeout to 5min.
We are experiencing test timeouts in the RHCOS Pipeline. Local testing shows
that 3min. seems to be around the time needed for the test.
Thus increasing the timeout from 3min. to 5min. should prevent future trouble.
The secex.ensure test does not timeout, but is on the border with 160s.
Increase timeout for it as well to prevent potential problems later.

Signed-off-by: Jan Schintag <jan.schintag@de.ibm.com>
(cherry picked from commit 81d97de70f3cdb3f9230c816ff7d33b3841dd964)