manifests: Inherit from Project sagano
Fedora CoreOS predates the time of bootable containers.  As such,
it's actually pretty large (relatively speaking).  Project Sagano
is a fresh take on more minimal base images.

At the same time, we want to de-duplicate efforts.  With this,
CoreOS cherry picks a few manifests from the Sagano "tier-0" and
"tier-1".

There's *definitely* more we can share between the two, but
this is a notable starting point.