manifests: Move kernel into separate manifest
This is prep for rebasing on [sagano](https://gitlab.com/cgwalters-playground/sagano)
where I hit on the issue that to build images with `kernel-rt`,
we need to clearly separate the `kernel` package from userspace
stuff (`systemd`, `rpm-ostree`) etc.

Note that we'll need to make the same change in RHCOS - until
we rebase both on sagano.