05core/dracut: enable `iscsi` module
In 9997df7b ("Move from `initramfs-args` in manifest to `dracut.conf.d`
files"), we migrated FCOS to use dracut dropins rather than CLI args.
One notable semantic change was that these dropins were also now
inherited by RHCOS.

One of the migrated directives was omitting the `iscsi` module. But RHCOS
OTOH has always included the `iscsi` dracut module, using `--add iscsi`.
This apparent conflict was noted in a comment and deemed safe[[1]].
And in fact, it *did* seem to work at the time, but only because there
was a bug in dracut: RHCOS happened to define some `--omit` arguments
as well and the bug made it so that the CLI arguments overrode the
inherited `omit_dracutmodules` dropin directives rather than being added
to them.[[2]]

The presence of this bug meant that `--add iscsi` still worked.
However, in RHEL9, the bug was fixed[[3]], which meant that we now
stopped pulling in the `iscsi` module since it's listed in the set
of modules to omit (`--add` arguments cannot undo earlier `--omit`/
`omit_dracutmodules` arguments).

The minimal fix needed here is to stop omitting the `iscsi` module. This
fixes https://issues.redhat.com/browse/OCPBUGS-19811, where some users
had already taken a dep on iSCSI iBFT support for autoconfiguration. It
also aligns with efforts to support root-on-iSCSI in the near future,
and there's no reason to not have that work happen in FCOS too.

A related larger task is to stop using `initramfs-args` in RHCOS since
it's now deprecated.[[4]]

[1]: https://github.com/coreos/fedora-coreos-config/pull/1828#discussion_r918220313
[2]: https://github.com/dracutdevs/dracut/issues/1341
[3]: https://github.com/dracutdevs/dracut/pull/1342
[4]: https://github.com/coreos/rpm-ostree/pull/3834