manifest: include crun-wasm and wasmedge-rt
We want to ship this in `next` for now to allow podman machines to make
use of it. We may either promote it to all streams (if there are more
people asking for it) or remove it at some point in the future (once
podman machines move to a layering model).

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/1375