tests/kernel-replace: extend timeout to 20m
We've seen this test timing out our ppc64le builder often lately.
It seems the heavy I/O for this test is causing that builder to
take longer and judging from the logs it looks like it is just getting
to the override of the kernel and is in the middle of the transaction
when it times out. Let't bump it to 20m.