05core: don't teardown network on switchroot for iSCSI boot
During an iSCSI boot, we rely on networking to access the root device.
So at no point can networking go down once the rootfs is mounted. This
means we have to skip the teardown logic in `coreos-teardown-initramfs`
in that case.

One way to look at this conceptually is: the goal of this service is to
reset everything so that the real root is configured on the first boot
like it would be on any other boot (i.e. avoid first boot-specific state
carried from the initramfs). Since initramfs networking is required on
every boot for iSCSI, it's true on every boot that networking needs to
be carried forward across switchroot. So the goal is still met since
it's not a first boot-only concern.

Part of: https://github.com/coreos/fedora-coreos-tracker/issues/1590