05core: make coreos-secex-ignition-decrypt.service virtio device dep explicit
The decryption script requires the encrypted Ignition virtio device to be
present. Previously, we used `Before=coreos-gpt-setup.service`, which I
think ran late enough that the device would always be present by then.
But we've now moved that service and `After=systemd-udevd.service` isn't
good enough so the service now fails because the device doesn't exist.

Fix this by directly listing the requirement on the device in
the unit, which is more explicit anyway. And because of systemd's
"ConditionPathExists don't turn off Requires" semantic, make the unit
dynamically enabled by the diskful generator. As a bonus, we now have
one less service in the default initramfs transaction (which actually
was getting pulled in on all arches too).