Add the candidate compose repo
This patch adds a repo definition for the latest candidate compose
for a given Fedora release. This is only relevant in the lead up to
Beta and GA for a given Fedora release where packages in the candidate
compose can lead the Fedora stable repos. For our `next` stream we want
to pick up the packages quickly, so using the candidate compose as input
makes sense.

Using this repo in the `next*` streams should address
https://github.com/coreos/fedora-coreos-tracker/issues/1602