denylist: disable run of ext.config.extensions.package
This is failing because the archive repo isn't active for F39 yet
and the version of VIM in the base layer is different than the versions
of packages that can be found in repos. Layering of vim packages is
disabled on next* streams for exactly this reason.

Adding it to the denylist for now. It will get removed in the next
promotion.