05core: order boot UUID restamp before ignition-kargs.service
It seems like the ext4 in el9 at least has an issue where the filesystem
can become corrupted if doing the UUID restamp at the same time as kargs
are being written to the BLS. It's better anyway to make these kinds of
invasive changes offline, so enforce that.

Follow-up to 5fb5b4fa ("05core: run `coreos-gpt-setup` and
`ignition-ostree-uuid-boot` later").