overlay.d: add 40grub overlay
bootupd upstream was modified to store default static grub configs
and allow for OS integrators to add their own snippets by placing
files in /usr/lib/bootupd/grub2-static/configs.d/ [1] When we move
to building with osbuild we will start taking advantage of this
functionality and calling bootupd with --with-static-configs. Let's
go ahead and add our snippets here so they can be leveraged when we
start calling bootupd with --with-static-configs.

The snippets added here come from the parts of the config currently
defined in coreos-assembler that weren't lifted into bootupd [2][3][4].

[1] https://github.com/coreos/bootupd/pull/543
[2] https://github.com/coreos/coreos-assembler/blob/4636b1a5c6dc00b1d6a58b1bfbb199431444336b/src/grub.cfg#L58-L60
[3] https://github.com/coreos/coreos-assembler/blob/4636b1a5c6dc00b1d6a58b1bfbb199431444336b/src/grub.cfg#L71-L85
[4] https://github.com/coreos/coreos-assembler/blob/4636b1a5c6dc00b1d6a58b1bfbb199431444336b/src/grub.cfg#L87-L91