kola/lvmdevices: use serial to reference disks
Using `/dev/vda`/`/dev/vdb` to refer to attached disks in inherently
unreliable since the order may change across e.g. OS or QEMU versions.

Slap on a serial name instead and refer to it that way. Interestingly,
it seems like the LVM tooling also picks up on this and bases its entry
in the `lvmdevices` file on the device serial.

Closes: #2807