tests/kernel-replace: tag with "reprovision"
This is a hack to get this test to run not in parallel with other
tests. The performance on our ppc64le builder has been abysmal and
this is one lever we can pull. We have bumped the timeout for the
test in the past. It's already at 20m. I say we just start running
this test serially.