05core: skip sysroot.mount for remote kdump
If kdump is setup to upload logs to a remote host, the
XFS module may not be loaded, which cause sysroot.mount to fail.

If the target for logs is SSH, kdump will boot with a `kdump_remote_ip`
karg. Skip generating sysroot.mount in that case so kdump can boot
without having to load XFS.

This is a temporary fix until we migrate older systems to have a root
karg: https://github.com/coreos/fedora-coreos-tracker/issues/1675

See https://issues.redhat.com/browse/OCPBUGS-27935