Enable composefs by default
We really want to aim to eventually enable this by default, let's
test things out in rawhide.

A thing that this is known to break is the "chattr -i" hack
for new toplevel dirs (xref https://github.com/coreos/rpm-ostree/issues/337 )

Basically if you want that, you either need to make a derived image,
or enable transient root.