Add `dnf5` on Fedora 41+
As per https://github.com/coreos/fedora-coreos-tracker/issues/1687,
we will for now add `dnf5` to the rawhide stream to test it out and
re-evaluate at branching.

It's not the default yet in f41, so doesn't own the `dnf` executable
name. Manually add a symlink to make playing with it more natural.