Misc: add a small tool to graph manifest includes
This is a simple script that display a tree of manifest includes.
I wrote it as a quick tool to get an idea of how manifest were
structured.

Requires the anytree lib.
Simply pass the manifest you want to start from as an argument.

Sample output on RHCOS:
```
(L)manifest.yaml
└── (L)manifest-rhel-coreos-9.yaml
    └── manifest-rhel-9.2.yaml
        ├── common.yaml
        │   ├── fedora-coreos-config/manifests/system-configuration.yaml
        │   ├── fedora-coreos-config/manifests/ignition-and-ostree.yaml
        │   │   └── fedora-coreos-config/manifests/bootable-rpm-ostree.yaml
        │   ├── fedora-coreos-config/manifests/file-transfer.yaml
        │   ├── fedora-coreos-config/manifests/networking-tools.yaml
        │   ├── fedora-coreos-config/manifests/user-experience.yaml
        │   ├── fedora-coreos-config/manifests/shared-workarounds.yaml
        │   └── packages-rhcos.yaml
        └── packages-openshift.yaml
```