overlays: drop `coreos-multipath-trigger.service`
This service was needed in the past to make multipath + LUKS work well.
The underlying bug seems to have been fixed now as I can no longer
reproduce it in Fedora or RHEL 9.4. Conveniently, this also works around
a bug in which that service would sometimes hang because of a bug[[1]]
in systemd which is still outstanding in RHEL 9.

Drop it.

We don't have any tests for this yet. Multipath + LUKS currently doesn't
work but should be fixed soon[[2]]. A test will be added as part of
that work.

[1]: https://github.com/systemd/systemd/issues/29863
[2]: https://github.com/coreos/fedora-coreos-tracker/issues/1728

Fixes: https://issues.redhat.com/browse/OCPBUGS-29325
(cherry picked from commit cc2e865b50bc2cf9d136b76ca6cf32b73757af4a)