kola/kernel-replace: adapt for RHCOS and SCOS
In one of the older issues, kola switch-kernel test was failing. In the discussion @jlebon suggested to remove named test and make it external, so that it can be utilized by FCOS, and also SCOS/RHCOS.
Additionally, the kernel version for FCOS was updated in this PR to the latest stable version `6.9.8-200.fc40`.

See: https://github.com/coreos/coreos-assembler/issues/1245