denylist: add kdump.crash for rawhide
We are trying to enable composeFS in rawhide and there is an issue
where kdump fails to generate the initrd from boot.
Manually trigerring the rebuild works but requires the extra manual
step.
Snoozing this test to let some time for the kdump team to
investigate.

Note that the kdump over SSH test works so we still have some
coverage for kdump.