tests/kernel-replace: use `dnf repoquery` to get kernel version
When testing from the c9s repos, don't hardcode a kernel version since
it'll eventually cycle out of the mirrors. Instead, just query the repo
to get the latest kernel.

Since we need the info we queried across reboots and we don't want to
query it each time (and possibly get different answers), persist it in
`/var` instead.

Ideally we'd do this for FCOS too, but we need to wait until the f41
rebase, which will bring dnf. (Or... could do it in a toolbox, but that
seems overkill.) But anyway for now, the kernel from the stable Fedora
repo shouldn't ever be GC'ed.