tests/podman/rootless-systemd: use the FCOS defined fedora.repo to set up container
Use the fedora.repo file defined in fedora-coreos-config to set up the
container. This will force packages to be downloaded from
dl.fedoraproject.org, as specified in the FCOS file. The ITUP cluster,
being used by the RHCOS pipeline, requires all outbound connections
to be specified in a Firewall Egress file, and this will ensure the
same connection will always be used.

Co-authored-by: Aashish Radhakrishnan <aaradhak@redhat.com>