manifests: add selinux-workaround.yaml for >= F41
Recent changes in the SELinux policy have broken a lot of our code.
Revert the affected domains back to permissive mode so we can
continue to build and test `releasever >= 41` until
https://github.com/fedora-selinux/selinux-policy/pull/2257 merges
and the domains are reverted upstream or until the issue is resolved
altogether.

Add the workaround for `afterburn_t` as well so we can unblock
https://github.com/coreos/fedora-coreos-tracker/issues/1784