tests/kola: use fedora-archive.repo for EOL Fedora containers
Use the fedora-archive.repo file defined in fedora-coreos-config
to set up the EOL containers in tests that use it. This will
force packages to be downloaded from `https://df.fedoraproject.org`,
as specified in the repo file. The ITUP cluster, being used by the
RHCOS pipeline, requires all outbound connections to be specified
in a Firewall Egress file, and this will ensure the same connection
will always be used when downloading the archived fedora content.

See: https://github.com/coreos/fedora-coreos-config/pull/3128