tests/upgrade: just reboot if any configuration changes were needed
I've seen nodes get into a stuck state if we made configuration changes
to the node in order to workaround the various starting state issues
over time. Let's just do a whole instance reboot if we made any
configuration changes to the instance to guarantee we don't end up
in a situation where the node is waiting forever to upgrade.