tests/kola: add more context on weird /var/mnt directory
On every boot, the `/var/mnt` directory *in the deployment root*
(i.e. not on top of any `/var` bind-mount/filesystem mount) would
get recreated and be unlabeled. After a lot of digging and busting out
systemtap, this turned out to be systemd doing this as part of switching
root as a temporary mount point.

In systemd v254+, this behaviour was changed to no longer require this
directory:

https://github.com/systemd/systemd/commit/f2c1d491a539035d6cc1fa53a7cef0cbc8d52902

For completeness, update the comment in this test to reflect these
findings.

(cherry picked from commit 53088cbfab1c5910d6e3cbce9eb5d73a7d93779e)