tests/upgrade: drop the rollback deployment on aarch64
We need to drop the rollback deployment. During upgrade
`...-> 40.20240906.1.0 (A)-> 41.20241109.1.0 (B)-> 42.20241114.91.0 (C)`
1) A->B, A has the unfixed ostree, the upgrade will copy dtb files
to `/boot/ostree` both for current A and new B with wrong label
2) B->C, B has the fixed ostree, the upgrade will prune A, leave B
with wrong label, and new C with correct label
3) Finaly booting C and will check that B has the wrong label
In this case we need to drop the rollback before checking.

If then upgrade to newer D, the upgrade will prune B (wrong label),
leave C with correct label, and new D with correct label. (We
should remove the drop under this case)