tests: fix boot.bootupd-validate on rhel-9.6
In the el-only variants, there is no `RHEL_VERSION` since the RHEL
version _is_ `VERSION_ID`.

Add a new `get_rhel_ver` commonlib function to abstract this away and
use that in the test.

Don't bother checking for the `VARIANT_ID`. We only care about the
coreos variant in the tests in this repo. (And once we start moving
tests to fedora-bootc, we would want to remove these checks anyway.)