tests/kola: rework derived kernel-replace
Now use dnf download to go ahead and download the RPMs that we'll
override replace in the container build.

Also, for Fedora, stop hardcoding the kernel to use and default to
using the latest kernel in the N-1 release of Fedora. It guarantees
it will be a different kernel and also is a moving target, which
means it won't get grotesquely out of date.