tests/kola: disable yum repos for kernel-replace test
Even though all the rpms that we are replacing are local rpm-ostree
will still reach out to the repos during the operation. Let's just
disable all the repos to prevent this from happening.

This was causing issues in our RHCOS pipeline because we have more
strict egress rules there and it was trying to hit the CentOS mirrors
when testing on c9s since the yum repos are there in c9s by default.

xref: https://github.com/openshift/os/issues/1673