overlay.d: drop 40grub overlay
The user.cfg was already dropped in f9a2278 and the Ignition
config we were providing here was only meant to be temporary
since Ignition is now shipping the required config as part
of the `ignition-grub` subpackage. Let's include the subpackage
and stop shipping any static configs here, which means we can
drop the overlay too.