tests: add coverage for multipath + multiple `boot` labels + no `root` karg
We want to be able to support the system having multiple filesystems
present with the `root` and `boot` labels.

This already mostly works today, because we add `root=UUID=` and
`boot=UUID=` kargs which disambiguate them. The exception is multipath,
where some places hardcode `/dev/disk/by-label/dm-mpath-[rb]oot` and
so don't make use of the UUIDs at all. This obviously then becomes
ambiguous in the presence of other similarly-labeled filesystems.

The previous patch plugged a large gap in this story.

Let's add a new test that covers the multipath + multiple boot disks +
no root kargs to ensure this keeps functioning.

Part of https://github.com/openshift/os/issues/1787.