manifests/shared-el: apply dracut fix in postcript
We are awaiting some dracut patches to land in Fedora mirror for [1].
We could wait for it, but we'd need it also for older rhcos branches
and those branches won't have the new dracut. So we decided to
carry part of the dracut-ng patch [2] in a post-script.
This will be backported in older rhcos branches, and will be reverted
on testing-devel once new dracut 107 available in mirror.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1937
[2] https://github.com/dracut-ng/dracut-ng/pull/1306