manifests: add `deriving` variable
This variable determines whether to include the minimal-plus manifests
at the treefile level. When `deriving` is false, it means that
we're doing a base rpm-ostree compose and we should include it. The
container-native flow will set it to true, so that the flattened
manifest only represents the things we want to add on top.