Build via Containerfile and derive from fedora-bootc
This builds Fedora CoreOS using a
Containerfile. It uses the [FROM scratch
flow](https://docs.fedoraproject.org/en-US/bootc/building-from-scratch/)
to do a base compose.

This is structured in a way that we can build FCOS using _both_ `podman
build` _and_ `cosa build`. This allows us to make the cutover much
smoother. So then, we could turn this on in e.g. rawhide first and let
it percolate down.

All the heavy lifting is done in the `build-rootfs` script. The idea is
that this script is shared by both FCOS and RHCOS. Random notes:
1. It's in Python. This is fine, because the builder image is separate
   from the target rootfs. The `bootc-base-imagectl` script itself is in
   Python and so can only be used from the `:standard` image.
2. Packages are extracted from the manifest, and fed to
   `bootc-base-imagectl` (see https://gitlab.com/fedora/bootc/base-images/-/merge_requests/178)
    so it's part of a single compose.
3. We keep respecting the in-git lockfiles. In the future, how lockfiles
   work will change (see discussions in
   https://github.com/coreos/fedora-coreos-tracker/issues/1861).
4. What you see overall is a lot of cosa-isms and rpm-ostree-isms being
   lifted and carried right into the build script. Things like overlays,
   versioning, postprocessing, os-release mutating, etc...

Rechunking is expected to be done as a secondary step. Once we have
https://github.com/containers/buildah/issues/5952, we can inline it back
into the Containerfile.